Notification Entity Done
