+++ lib/src/app/domain/entities/content.dart
@@ -1,0 +1,8 @@
+import 'dart:js';
+
+class Content {
+  String? _content;
+
+  String? getContent() => _content;
+  void setContent(String value) => _content = value;
+}
