+++ lib/src/app/domain/entities/notification.dart
@@ -1,5 +1,5 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
-import 'notification_properties.dart';
+import 'content.dart';
 
 class Notification {
    String? _notificationId;
@@ -16,3 +16,22 @@ class Notification {
   void setCanceledAt(DateTime value) => _properties?.canceledAt = value;
 
 }
+
+class NotificationProperties {
+  String? recipientId;
+  String? content;
+  String? category;
+  DateTime? readAt;
+  DateTime? canceledAt;
+  DateTime? createdAt;
+
+  NotificationProperties({
+    this.recipientId,
+    this.content,
+    this.category,
+    this.readAt,
+    this.canceledAt,
+    this.createdAt,
+  });
+}
+
